Fix direct metrics smoke link
