+++ native/zero-c/src/target_backend.c
@@ -140,24 +140,6 @@ size_t z_direct_backend_symbol_overhead(ZDirectBackend backend, bool has_readonl
   return 0;
 }
 
-size_t z_direct_target_stack_bytes(const ZTargetInfo *target, const IrProgram *program) {
-  if (!program) return 0;
-  if (z_direct_object_backend(target) == Z_DIRECT_BACKEND_MACHO64 ||
-      z_direct_exe_backend(target) == Z_DIRECT_BACKEND_MACHO64) {
-    return z_macho64_stack_bytes_from_ir(program);
-  }
-  return program->direct_stack_bytes;
-}
-
-size_t z_direct_target_max_frame_bytes(const ZTargetInfo *target, const IrProgram *program) {
-  if (!program) return 0;
-  if (z_direct_object_backend(target) == Z_DIRECT_BACKEND_MACHO64 ||
-      z_direct_exe_backend(target) == Z_DIRECT_BACKEND_MACHO64) {
-    return z_macho64_max_frame_bytes_from_ir(program);
-  }
-  return program->direct_max_frame_bytes;
-}
-
 bool z_direct_backend_supports_runtime_object(ZDirectBackend backend) {
   const ZDirectBackendDescriptor *descriptor = direct_backend_descriptor(backend);
   return descriptor ? descriptor->runtime_object_supported : false;
